+++ src/main/java/roomescape/reservation/dao/JdbcReservationTimeDao.java
@@ -10,6 +10,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.reservation.domain.ReservationTime;
+import roomescape.reservation.service.ReservationTimeDao;
 
 @Repository
 public class JdbcReservationTimeDao implements ReservationTimeDao {
