+++ src/main/java/roomescape/reservation/dao/JdbcThemeDao.java
@@ -11,6 +11,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.reservation.domain.Theme;
+import roomescape.reservation.service.ThemeDao;
 
 @Repository
 public class JdbcThemeDao implements ThemeDao {
